+++ src/VM/VM.cpp
@@ -109,9 +109,10 @@ namespace Ark
         std::string path = "./" + file;
 
         if (m_state->m_filename != ARK_NO_NAME_FILE)  // bytecode loaded from file
-            path = "./" + (fs::path(m_state->m_filename).parent_path() / fs::path(file)).string();
+            path = (fs::path(m_state->m_filename).parent_path() / fs::path(file)).relative_path().string();
         std::string lib_path = (fs::path(m_state->m_libdir) / fs::path(file)).string();
 
+        // if it's already loaded don't do anything
         if (std::find_if(m_shared_lib_objects.begin(), m_shared_lib_objects.end(), [&, this](const auto& val) {
             return (val->path() == path || val->path() == lib_path);
         }) != m_shared_lib_objects.end())
@@ -122,7 +123,7 @@ namespace Ark
         else if (Utils::fileExists(lib_path))  // check in LOAD_PATH otherwise
             m_shared_lib_objects.emplace_back(std::make_shared<SharedLibrary>(lib_path));
         else
-            throwVMError("could not load plugin: " + file);
+            throwVMError("Could not find plugin '" + file + "'. Searched in\n\t- " + path + "\n\t- " + lib_path);
 
         // load data from it
         mapping* map;
@@ -130,7 +131,7 @@ namespace Ark
         try {
             map = m_shared_lib_objects.back()->template get<mapping* (*)()>("getFunctionsMapping")();
         } catch (const std::system_error& e) {
-            throwVMError(std::string(e.what()));
+            throwVMError("An error occurred while loading plugin '" + file + "': " + std::string(e.what()));
         }
 
         std::size_t i = 0;